progress on snapcast
